Handle CTRL+C within the console -> clear input buffer
This feels more like other known interpreters.

Threaded mode is kinda screwed up because PyThreadState_SetAsyncExc
doesn't cause byte-code execution in the thread and therefore doesn't
wake up the thread from neither the Condition variable wait nor C calls.

The condition part can be mitigated by calling stdin.flush(), but there
seems to be no way to cause a signal handler to run in the thread.

However, this is not due to this patch but was the same already on
master.