Improve insertFromMimeData behaviour to paste, not evaluate!
This fixes a long standing issue with the usage of evaluate_buffer for
pasting. This approach is unintuitive in that it immediately evaluated
the pasted text instead of simply inserting it into the current buffer.

The handling is now still not ideal for multi-line strings, but this
issue will be addressed in a later commit.