Feed lines directly to interpreter via signals
While using a "stream"-like stdin/stdout object to communicate the state
of the interpreter seems nice and very general at first, I came to the
conclusion that it is in fact a bit too general and holds back the full
potential of this GUI console in some regards, which is more than a text
based REPL. In particular:

- it makes it harder to implement `raw_input` to be usable for prompts
  during code execution

- to improve multiline handling, copy-pasting, and handling of the
  in/out/more prompts, the console should have full control over which
  regions are input and output rather than treating them equally

This means eventually moving all UI management responsibilities fully to
the console and leave the interpreter for code execution only. Once this
is done, it will become even clearer that there is no need for using a
stream-like object to push code.