Hide the cursor during the execution of a python command
This has the nice side effect of showing the prompt as well.

Without this fix the prompt was not shown after a command that took
more than a millisecond to execute.
This can be illustrated by running time.sleep(1)
With this fix the cursor is hidden for 1 second (indicating that the system
is busy), after the second the next prompt appears and the cursor reappears.

Do not accept input when a command is executing.

The only exception is Ctrl-C, but the ctrl_c handler does not seem to
be doing anything in practice.