Close the thread in the hideEvent() callback and simplify exit()
Doing the thread closure in hideEvent() ensure that the thread is
stopped when you close the window via the window manager and using
exit() in the interpreter.
Using the immediate close makes the widget more compatible when embedded
in another application. The original code would close the entire app
instead of just closing the interpreter widget.